Notice: Updating to this version will reset the positions of your FX frames!!!
Converted the Forte module over to the new DB
   - addons/forte.lua
   - modules/themes.lua
   - LUI.lua

Fixed issue with Blizzard's Floating combat text not changing to the set font.
   - LUI.lua

Fixed issue with tooltip showing under the framelevel of the friends/guild list.
   - modules/datatext.lua


Changed filename of extras.toc.txt to LUI_ExtraModule.toc.txt (this is used as a template for the toc file of third-party modules).