- InfoText Guild: hopefully fix the taint issues caused by WoW 5.4.1, however unfortunately I've had to simplify the Guild OnClick.
It will now only open the main Guild page, it will not open to the roster. I will keep looking for a fix for this but it seems that the "GuildFrameTabX:Click()" calls were causing the issue. Once removed it all worked again.

Hopefully this is the last of the WoW 5.4.1 issues that need an urgent fix. *crosses fingers*